Merge master into the gasp-graph-isolation branch to prepare PR #361
for a clean squash merge. Resolved the single expected conflict in
pkg/server/internal/app/lookup_question_service_test.go by taking
master's formatting verbatim (struct literal line-breaking only, no
semantic difference). go.mod/go.sum and
pkg/server/internal/app/lookup_question_service.go auto-merged to
master's content.

+++ pkg/server/internal/ports/openapi/openapi_non_admin_response_types.gen.go
@@ -59,8 +59,9 @@ type Metadata map[string]ServiceMetadata
 
 // OutputListItem defines model for OutputListItem.
 type OutputListItem struct {
-	Beef        []byte `json:"beef"`
-	OutputIndex uint32 `json:"outputIndex"`
+	// Beef BEEF bytes encoded as the BRC-24 JSON byte array. Earlier Go server releases emitted a base64 string here through Go's []byte JSON encoding; that legacy representation is not BRC-24 JSON interoperable.
+	Beef        []int32 `json:"beef"`
+	OutputIndex uint32  `json:"outputIndex"`
 }
 
 // RequestSyncRes defines model for RequestSyncRes.
